feat(TpFilmotheque): replace long id type by int + add update genre + tests

+++ spring/tp/TpFilmotheque/src/main/java/fr/eni/tp/filmotheque/bll/mock/FilmServiceBouchon.java
@@ -43,7 +43,7 @@ public class FilmServiceBouchon implements FilmService {
 	 *//*
 
 	@Override
-	public Film consulterFilmParId(long id) {
+	public Film consulterFilmParId(int id) {
 		return lstFilms.stream().filter(item -> item.getId() == id).findAny().orElse(null);
 	}
 
@@ -58,7 +58,7 @@ public class FilmServiceBouchon implements FilmService {
 	}
 
 	@Override
-	public Genre consulterGenreParId(long id) {
+	public Genre consulterGenreParId(int id) {
 		return lstGenres.stream().filter(item -> item.getId() == id).findAny().orElse(null);
 	}
 
@@ -69,7 +69,7 @@ public class FilmServiceBouchon implements FilmService {
 	 *//*
 
 	@Override
-	public Participant consulterParticipantParId(long id) {
+	public Participant consulterParticipantParId(int id) {
 		return lstParticipants.stream().filter(item -> item.getId() == id).findAny().orElse(null);
 	}
