+++ spring/tp/TpFilmotheque/src/main/java/fr/eni/tp/filmotheque/bo/Participant.java
@@ -1,7 +1,7 @@
 package fr.eni.tp.filmotheque.bo;
 
 public class Participant extends Personne{
-    public Participant(long id, String nom, String prenom) {
+    public Participant(int id, String nom, String prenom) {
         super(id, nom, prenom);
     }
